Handle shopify sending paginated responses
