+++ server/utils.js
@@ -1,27 +1,30 @@
-const createPoll = (fn, interval) => {
-  const subscribers = {};
-  const intervals = {};
-  return ({ getKey, args, onSuccess }) => {
-    const key = getKey();
-    if (!subscribers[key]) {
-      subscribers[key] = [];
-    }
-    const index = subscribers[key].push(onSuccess) - 1;
-    if (!intervals[key]) {
-      intervals[key] = setInterval(() => {
-        if (subscribers[key].length < 1) {
-          clearInterval(intervals[key]);
-          intervals[key] = null;
-        }
-        fn(...args).then((...args) =>
-          subscribers[key].forEach(s => s(...args))
-        );
-      }, interval);
-    }
-    return () => {
-      subscribers[key].splice(index, 1);
-    };
-  };
+const getToday = () => {
+  const date = new Date();
+  date.setHours(0, 0, 0, 0);
+  return date.toISOString();
 };
 
-module.exports = { createPoll };
+const getYesterday = () => {
+  const date = new Date();
+  date.setHours(0, 0, 0, 0);
+  date.setDate(date.getDate() - 1);
+  return date.toISOString();
+};
+
+const getDateFromString = str =>
+  ({
+    today: getToday(),
+    yesterday: getYesterday()
+  }[str]);
+
+const paginate = async (params, request) => {
+  let responses = [];
+  do {
+    const response = await request(params);
+    responses = [...responses, ...response];
+    params = response.nextPageParameters;
+  } while (params !== undefined);
+  return responses;
+};
+
+module.exports = { getToday, getYesterday, getDateFromString, paginate };
